requirements.txt
absl-py>=0.13.0
certifi>=2021.5.30
chardet>=4.0.0
flatbuffers>=2.0
idna>=2.10
jax>=0.2.16
jaxlib>=0.1.68
joblib>=1.0.1
liac-arff>=2.5.0
minio>=7.0.4
numpy>=1.21.0
openml>=0.12.2
opt-einsum>=3.3.0
pandas>=1.2.5
pyarrow>=4.0.1
python-dateutil>=2.8.1
pytz>=2021.1
requests>=2.25.1
scikit-learn>=0.24.2
scipy>=1.7.0
six>=1.16.0
sklearn>=0.0
threadpoolctl>=2.1.0
urllib3>=1.26.6
xmltodict>=0.12.0
tabulate>=0.8.9
termcolor>=1.1.0
